requirements.txt
pytesseract==0.2.6
numpy==1.14.0
regex==2018.02.08
pyyaml==3.12
opencv-python==4.5.5.64
opencv-contrib-python==4.5.5.64
